Add task status tracking capability to Duke
Duke can now mark tasks as done, and reflect the status of the task (done or not done) whilst printing the list of tasks. Created a new class 'Task' for better encapsulation. Text encoding for 'Task Status Icon' fishy in my machine (IDE prints question marks instead of unicode characters), but the logic of the code is present.