Level 10. GUI
Changes made in this commit:
1) Implemented GUI using Javafx library
2) Updated Parser such that it checks for valid date format for Events

fxml files and images were added to a newly created resources folder
to ensure the functioning of the javafx UI. As output must be constantly
managed as String in the dialog boxes, major refactoring has been made
to the Ui class such that every method that used to print a message to
the system now appends the message to a StringBuilder object before
being extracted as a String.

To be aligned with the Duke requirement, Event tasks can only accept
a specific Date format for the additional info like Deadline tasks.

Other notes:
The tick and cross symbols, along with the emoticon symbol found in
most warning messages, have been replaced with their unicode form as
their previous forms cannot be printed on the javafx ui.