Add JavaFX functionality.
Edit build.gradle to add JavaFX functionality.

Duke's constructor had to be changed to take in no arguments. Previously, a String representing the saveFilePath was passed into the constructor, but this had to be changed to accommodate the JavaFX launch code. Now, the saveFilePath can still be edited, but any change has to be made directly to the static variable, instead of being passed to the Duke constructor directly. An example is given in the main() method of Duke.