Fix save/load feature.
Previously task list save file would be read from and written to resource directory,
this causes Duke to be unable to save task list when run from jar as the jar is locked.

Change so that a new directory in the directory the jar is contained in to be created,
and save files read to and written from there.