Refactor: make Ui more flexible
Previously, the execute method of Commands would return a CommandResult
object and the call site is responsible for displaying the result. This
is awkward because commands have different result formats. Ideally
CommandResults should be subclassed to account for the different
formats, but this results in too many small classes. For simplicity, I
stored all types of results (together with their markup) as Strings and
printed them regardless of whether the current Ui is CLI or GUI.

Instead of returning a CommandResult, the execute method of Command now
takes in a Ui and calls the appropriate methods on it to interact with
the user. Both Cli and Gui implement Ui, so they are free to
define how they want to display the information to the user. For
example, Cli might display plaintext while MainWindow might prefer to
wrap the information in GUI elements. Currently these methods are
defined as defaults in the Ui interface (and they only print plaintext).
However, they can be overriden to provide more advanced Ui that are not
previously possible with packaged information from CommandResult.